Corrections Ax12

+++ src/Driver-EV3/AsservDriver.cpp
@@ -2,15 +2,14 @@
 
 #include "AsservDriver.hpp"
 
-#include <cerrno>
-#include <cstring>
+#include <string>
 
 #include "../Log/Logger.hpp"
 
 using namespace std;
 using namespace ev3dev;
 
-AAsservDriver * AAsservDriver::create(std::string )
+AAsservDriver * AAsservDriver::create(std::string)
 {
 	static AsservDriver *instance = new AsservDriver();
 	return instance;
@@ -18,11 +17,9 @@ AAsservDriver * AAsservDriver::create(std::string )
 
 void AsservDriver::reset()
 {
-	if (_motor_left.connected())
-		_motor_left.reset();
+	if (_motor_left.connected()) _motor_left.reset();
 
-	if (_motor_right.connected())
-		_motor_right.reset();
+	if (_motor_right.connected()) _motor_right.reset();
 }
 
 AsservDriver::AsservDriver()
@@ -33,32 +30,28 @@ AsservDriver::AsservDriver()
 	if (mright.connected())
 	{
 		_motor_right = OUTPUT_D;
-		logger().info() << "EV3 Motor (RIGHT) - " << mright.address()
-				<< " connected ;CPR=" << mright.count_per_rot()
-				<< " Name=" << mright.driver_name() << " Pol="
+		logger().info() << "EV3 Motor (RIGHT) - " << mright.address() << " connected ;CPR="
+				<< mright.count_per_rot() << " Name=" << mright.driver_name() << " Pol="
 				<< mright.polarity() << logs::end;
 	}
 	else
 	{
 		_motor_right = OUTPUT_D;
-		logger().error() << "ERROR OUTPUT_A - Motor (RIGHT) "
-				<< "not connected !!" << logs::end;
+		logger().error() << "ERROR OUTPUT_A - Motor (RIGHT) " << "not connected !!" << logs::end;
 	}
 
 	motor mleft = motor(OUTPUT_A);
 	if (mleft.connected())
 	{
 		_motor_left = OUTPUT_A;
-		logger().info() << "EV3 Motor (LEFT) - " << mleft.address()
-				<< " connected ;CPR=" << mleft.count_per_rot()
-				<< " Name=" << mleft.driver_name() << " Pol="
+		logger().info() << "EV3 Motor (LEFT) - " << mleft.address() << " connected ;CPR="
+				<< mleft.count_per_rot() << " Name=" << mleft.driver_name() << " Pol="
 				<< mleft.polarity() << logs::end;
 	}
 	else
 	{
 		_motor_left = OUTPUT_A;
-		logger().error() << "ERROR OUTPUT_D - Motor (LEFT) "
-				<< "not connected !!" << logs::end;
+		logger().error() << "ERROR OUTPUT_D - Motor (LEFT) " << "not connected !!" << logs::end;
 	}
 
 	if (_motor_right.connected()) //if both motors are connected, then initialize each motor.
@@ -137,8 +130,8 @@ void AsservDriver::setMotorLeftPower(int power, int timems)
 			_motor_left.set_speed_sp(power);
 
 			logger().debug() << "LEFT current motor speed in ticks per second:"
-					<< _motor_left.speed() << " count_per_rot:"
-					<< _motor_left.count_per_rot() << logs::end;
+					<< _motor_left.speed() << " count_per_rot:" << _motor_left.count_per_rot()
+					<< logs::end;
 		}
 		else if (_motor_left.speed_regulation_enabled() == "off") //duty_cycle_sp
 		{
@@ -171,14 +164,14 @@ int AsservDriver::limit(int power, int max)
 {
 	if ((power < -max))
 	{
-		logger().info() << "ERROR Motor power " << power << " exceeded minimum "
-				<< -max << "!!" << logs::end;
+		logger().info() << "ERROR Motor power " << power << " exceeded minimum " << -max << "!!"
+				<< logs::end;
 		power = -max;
 	}
 	else if (power > max)
 	{
-		logger().info() << "ERROR Motor power " << power << "exceeded maximum "
-				<< max << "!!" << logs::end;
+		logger().info() << "ERROR Motor power " << power << "exceeded maximum " << max << "!!"
+				<< logs::end;
 		power = max;
 	}
 	return power;
@@ -199,8 +192,8 @@ void AsservDriver::setMotorRightPower(int power, int timems)
 			_motor_right.set_speed_sp(power);
 
 			logger().debug() << "RIGHT current motor speed in ticks per second:"
-					<< _motor_right.speed() << " count_per_rot:"
-					<< _motor_right.count_per_rot() << logs::end;
+					<< _motor_right.speed() << " count_per_rot:" << _motor_right.count_per_rot()
+					<< logs::end;
 		}
 		else if (_motor_right.speed_regulation_enabled() == "off") //duty_cycle_sp
 		{
@@ -243,7 +236,7 @@ long AsservDriver::getLeftInternalEncoder()
 	if (_motor_left.connected())
 	{
 		//+/- 2,147,483,648
-		return  _motor_left.position();
+		return _motor_left.position();
 	}
 	else
 		return 0;
@@ -253,7 +246,7 @@ long AsservDriver::getRightInternalEncoder()
 {
 	if (_motor_right.connected())
 	{
-		return  _motor_right.position();
+		return _motor_right.position();
 	}
 	else
 		return 0;
@@ -340,32 +333,29 @@ void AsservDriver::enableRightHardRegulation(bool enable)
 		}
 	}
 }
-
-void AsservDriver::enableHardRegulation(bool enable)
-{
-	enableLeftHardRegulation(enable);
-	enableRightHardRegulation(enable);
-}
-
-
-
-
+/*
+ void AsservDriver::enableHardRegulation(bool enable)
+ {
+ enableLeftHardRegulation(enable);
+ enableRightHardRegulation(enable);
+ }
+ */
 
 /*
-float AsservDriver::odo_GetX_mm()
-{
-	return 0.0;
-}
-float AsservDriver::odo_GetY_mm()
-{
-}
-float AsservDriver::odo_GetTheta_Rad()
-{
-}
-float AsservDriver::odo_GetTheta_Degree()
-{
-}*/
-void AsservDriver::odo_SetPosition(double x_m, double y_m, float angle_rad)
+ float AsservDriver::odo_GetX_mm()
+ {
+ return 0.0;
+ }
+ float AsservDriver::odo_GetY_mm()
+ {
+ }
+ float AsservDriver::odo_GetTheta_Rad()
+ {
+ }
+ float AsservDriver::odo_GetTheta_Degree()
+ {
+ }*/
+void AsservDriver::odo_SetPosition(double x_m, double y_m, double angle_rad)
 {
 }
 RobotPosition AsservDriver::odo_GetPosition()
@@ -413,5 +403,4 @@ void AsservDriver::motion_AssistedHandling()
 void AsservDriver::motion_ActivateManager(bool enable)
 {
 }
-
